[OSSIE][NVIDIA_GSF] Rename GSF converter directory and package to NVIDIA
Name the converter after the vendor that owns it rather than the product,
matching how every other converter directory in the tree is named. The
directory becomes converters/nvidia and the import package becomes
ossie_nvidia_gsf, so the product name still appears at the point of import.

Rename the distribution to apache-ossie-nvidia-gsf and the console script to
ossie-nvidia-gsf, and repoint the CI workflow, itself renamed to
converter-nvidia-ci.yml to preserve the converter-<dir>-ci.yml convention
shared by the other converters. Its check name changes accordingly.

The NVIDIA_GSF vendor code in the spec's extension table is left alone, being
a format identifier rather than a path.
