+++ converters/nvidia/pyproject.toml
@@ -26,7 +26,7 @@ dev = [
 ]
 
 [project]
-name = "apache-ossie-gsf"
+name = "apache-ossie-nvidia-gsf"
 version = "0.1.0.dev0"
 description = "NVIDIA GSF <> Apache Ossie offline YAML converter"
 authors = [{ name = "Apache Software Foundation", email = "dev@ossie.apache.org" }]
@@ -46,14 +46,14 @@ dependencies = [
 ]
 
 [project.scripts]
-ossie-gsf = "ossie_gsf.converter:main"
+ossie-nvidia-gsf = "ossie_nvidia_gsf.converter:main"
 
 [project.urls]
 homepage = "https://ossie.apache.org/"
 repository = "https://github.com/apache/ossie/"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/ossie_gsf"]
+packages = ["src/ossie_nvidia_gsf"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
